Refactor specs for metric implementations
